added means to resolve root body and system
in order to validate the license of an object, we need to try
to resolve which body or system the object originates from
thus i added an abstract method to OParl.Object which does
so. in some places it can not yet be guaranteed that every
object resolves to a valid body. we have to implement means to
cover an objects incompleteness. for example when it has been
loaded as the internal object of another object. in this case
we have to make sure that the object is fully loaded and contains
all neccessary backreferences.