[wip] resolving fields are now methods
[!] neither tests nor docs will build from this commit!
there are fields that resolve to other objects when they
are being getted by the user. on this occasion there's the
possibility that parsing the underlying object fails which
liboparl will express through a ParsingError. We cannot throw
or escalate an Exception inside a getter of a property, so
all properties that resolve other objects have been transformed
into methods