solve memory-corruption problem
when using the validation method trough pyGI, it was a problem
to handle the list to the python-side. now it's being passed
as an unowned reference. the validated object holds a reference
to its own results.
this could be somewhat expensive in terms of memory but as a first
solution it works well enough. If you want to use the returned list
even when the object will probably die, it is recommended to create
a copy