Created EmbeddedObject for easy object completion
Some objects can be embedded into other objects.
Unfortunately for us, these embededded versions do not
necessarily contain all backreferences.

This commit introduces a method for autoloading (refreshing)
an object on requesting a non-resolved backreference if the
object was not yet resolved. A simpler approach would have
been to always reload. That approach was disregarded in favor
of saving on traffic. The latter is also the reason why the
autoload is not simply applied on object load but in the
actual backreference methods since doing otherwise would
loose the benefits of embedded objects.