Simplify indexable getter and iterator methods
PageableSequence.get():

Implements nullable getting of objects with automatic refetching
if possible. Please be aware that this method has a best case
complexity of O(1) => Object is available and a worst case complexity
of O(n) if new objects need to be fetched and parsed.

PageableSequence.next_value():

Provides iteration functionality to use PageableSequences in
`for x in s`-constructs.