Fix PageableSequence.get()
The condition for fetching not-currently-available objects was
not quite correct. Before it was impossible to ask for object
indexes not on the currently loaded pages or on the next page
to load. This is now fixed by iterating over the pager loader
until the object is found.

This may result in never finding any object and while determining
that loading all objects of a list into memory. There is room
for improvement here.