Change versioning to just use the tag's value
Since naming requirements for GI are sorta strict, it's
much more feasable to just stick to getting the version
from git-describe instead of appending the current commit
hash.

This does not enhance versioning as much as I would have liked
but at least removes the need to track the library version
inside the meson file as well as in the git tags.

There is still some work to do as this should actually be split
into the version components and correctly set `api` and
`revision` as well as probably dump that into a vala-includable
file which then can be used to provide a `OParl.library_version()`
or something like that to read the version at runtime.