Print the object url on error
The "Tried to make an object from a json without type" is the error I've seen most frequently. It is thrown whenever the type is really missing, but also when the server returned an error instead of the actual object. One can not say though on which object the error occured which is quite a problem for debugging, so this change adds this information.